+++ client/src/features/results/Navigation.tsx
@@ -118,7 +118,7 @@ const Navigation: React.FC<INavigation> = ({
               {Object.entries(searchScope).map(([key, value]) => (
                 <StyledNavLink
                   key={key}
-                  to={`/view/results/${key}?${
+                  to={`/view/results/${key}${key === 'collections' ? '/all' : ''}?${
                     (advancedSearch && !urlParams.has('qt') && key !== qt) ||
                     isSwitchToSimpleSearch
                       ? `${urlParams.toString()}&qt=${tab}`