+++ client/src/features/results/ResultsPage.tsx
@@ -37,6 +37,7 @@ const ResponsiveCol = styled(Col)`
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const getScopedResultsComponent: any = (
   tab: string,
+  subTab: string | undefined,
   searchResponse: ISearchResponse,
   isMobile: boolean,
 ) => {
@@ -48,7 +49,13 @@ const getScopedResultsComponent: any = (
     case 'works':
       return <WorkResults searchResponse={searchResponse} isMobile={isMobile} />
     case 'collections':
-      return <SetResults searchResponse={searchResponse} isMobile={isMobile} />
+      return (
+        <SetResults
+          searchResponse={searchResponse}
+          isMobile={isMobile}
+          nestedPage={subTab}
+        />
+      )
     case 'people':
       return (
         <PersonResults searchResponse={searchResponse} isMobile={isMobile} />
@@ -73,7 +80,7 @@ const title = 'Results Page'
 
 const ResultsPage: React.FC = () => {
   const dispatch = useAppDispatch()
-  const { tab } = useParams<keyof ResultsTab>() as ResultsTab
+  const { tab, subTab } = useParams<keyof ResultsTab>() as ResultsTab
   const paramPrefix = getParamPrefix(tab)
   const [isMobile, setIsMobile] = useState<boolean>(
     window.innerWidth < theme.breakpoints.md,
@@ -184,7 +191,7 @@ const ResultsPage: React.FC = () => {
           </Col>
         ) : (
           <Col xs={12} className={isMobile ? '' : 'px-0'}>
-            {getScopedResultsComponent(tab, searchResponse, isMobile)}
+            {getScopedResultsComponent(tab, subTab, searchResponse, isMobile)}
           </Col>
         )}
       </StyledEntityPageSection>